improv

+++ core/utils/27-litegapps.sh
@@ -3,26 +3,26 @@
 # Litegapps addon.d (running in rom installer)
 # by wahyu6070
 
+. /tmp/backuptool.functions
 log=/data/media/0/Android/litegapps/litegapps_addon.d.log
 base=/data/kopi/modules/litegapps
 
 test ! -d $(dirname $log) && mkdir -p $(dirname $log)
-if [ -f $log ] && [ $(du -sk $log | cut -f1) -eq 1000 ]; then
-	rm -rf $log
-fi
+test -f $log && rm -rf $log
+
 ps | grep zygote | grep -v grep >/dev/null && BOOTMODE=true || BOOTMODE=false
 $BOOTMODE || ps -A 2>/dev/null | grep zygote | grep -v grep >/dev/null && BOOTMODE=true
 
 if ! $BOOTMODE; then
 # update-binary|updater <RECOVERY_API_VERSION> <OUTFD> <ZIPFILE>
-OUTFD=$(ps | grep -v 'grep' | grep -oE 'update(.*) 3 [0-9]+' | cut -d" " -f3)
-[ -z $OUTFD ] && OUTFD=$(ps -Af | grep -v 'grep' | grep -oE 'update(.*) 3 [0-9]+' | cut -d" " -f3)
-# update_engine_sideload --payload=file://<ZIPFILE> --offset=<OFFSET> --headers=<HEADERS> --status_fd=<OUTFD>
-[ -z $OUTFD ] && OUTFD=$(ps | grep -v 'grep' | grep -oE 'status_fd=[0-9]+' | cut -d= -f2)
-[ -z $OUTFD ] && OUTFD=$(ps -Af | grep -v 'grep' | grep -oE 'status_fd=[0-9]+' | cut -d= -f2)
-fi
+ OUTFD=$(ps | grep -v 'grep' | grep -oE 'update(.*) 3 [0-9]+' | cut -d" " -f3)
+ [ -z $OUTFD ] && OUTFD=$(ps -Af | grep -v 'grep' | grep -oE 'update(.*) 3 [0-9]+' | cut -d" " -f3)
+ # update_engine_sideload --payload=file://<ZIPFILE> --offset=<OFFSET> --headers=<HEADERS> --status_fd=<OUTFD>
+ [ -z $OUTFD ] && OUTFD=$(ps | grep -v 'grep' | grep -oE 'status_fd=[0-9]+' | cut -d= -f2)
+ [ -z $OUTFD ] && OUTFD=$(ps -Af | grep -v 'grep' | grep -oE 'status_fd=[0-9]+' | cut -d= -f2)
+ fi
+ ui_print() { $BOOTMODE && log -t Magisk -- "$1" || echo -e "ui_print $1\nui_print" >> /proc/self/fd/$OUTFD; }
 
-ui_print() { $BOOTMODE && log -t Magisk -- "$1" || echo -e "ui_print $1\nui_print" >> /proc/self/fd/$OUTFD; }
 print(){
 	ui_print "$1"
 	}
@@ -33,42 +33,8 @@ sedlog(){
 ch_con(){
 chcon -h u:object_r:system_file:s0 "$1" || sedlog "Failed chcon $1"
 }
-COPY_FILE() {
-	local src=$1
-	local dest=$2
-	cp -dpf "$src" "$dest"
-  # symlinks don't have a context
-  if [ ! -L "$src" ]; then
-    # it is assumed that every label starts with 'u:object_r' and has no white-spaces
-    local context=`ls -Z "$src" | grep -o 'u:object_r:[^ ]*' | head -1`
-    chcon "$context" "$dest"
-  fi
-}
-BACKUP_FILE(){
-	local src=$1
-	local DIR_NAME=`dirname $src`
-	local DIR_BACKUP=$DIR${DIR_NAME}
-	print "$src"
-	if [ -f $src ] && [ ! -L $src ]; then
-		test ! -d $DIR_BACKUP && mkdir -p $DIR_BACKUP
-		COPY_FILE $src $DIR${src}
-	fi
-	
-	}
-RESTORE_FILE(){
-	local FILE=$DIR${1}
-	local DEST=$1
-	if [ -f $FILE ] && [ ! -L $FILE ]; then
-		local DEST_DIR=`dirname $DEST`
-		if [ ! -d $DEST_DIR ]; then
-			mkdir -p $DEST_DIR
-			chmod 755 $DEST_DIR
-		fi
-		COPY_FILE $FILE $DEST
-		#rm -rf $FILE
-	fi
-	
-	}
+
+
 if [ -f /system/system/build.prop ]; then
 SYSTEM=/system/system
 elif [ -f /system_root/system/build.prop ]; then
@@ -93,18 +59,16 @@ fi
 
 [ ! -d $base ] && return 0
 
-# backup/restore directory
-DIR=/dev/litegapps
-test ! -d $DIR && mkdir -p $DIR
+# S = is variable in backup/restore script in rom flashable
 
 case "$1" in
   backup)
   	print "Backuping LiteGapps"
   	if [ -f $base/list_install_system ]; then
   		for A in $(cat $base/list_install_system); do
-  			if [ -f $STSTEM/$A ] && [ ! -L $SYSTEM/$A ] ; then
-  				sedlog "  Backuping •> $SYSTEM/$A"
-  				BACKUP_FILE $SYSTEM/$A
+  			if [ -f $S/$A ] && [ ! -L $S/$A ] ; then
+  				sedlog "  Backuping •> $S/$A"
+  				backup_file $S/$A
   			fi
     	  done
  	 fi
@@ -112,7 +76,7 @@ case "$1" in
  	 	for B in $(cat $base/list_install_product); do
  	 		if [ -f $PRODUCT/$B ] && [ ! -L $PRODUCT/$B ] ; then
  	 			sedlog "  Backuping •> $PRODUCT/$B"
- 	 			BACKUP_FILE $PRODUCT/$B
+ 	 			backup_file $PRODUCT/$B
     		  fi
     	  done
   	fi
@@ -120,7 +84,7 @@ case "$1" in
   		for C in $(cat $base/list_install_system_ext); do
   			if [ -f $SYSTEM_EXT/$C ] && [ ! -L $SYSTEM_EXT/$C ] ; then
   				sedlog "  Backuping •> $SYSTEM_EXT/$C"
-  				BACKUP_FILE $SYSTEM_EXT/$C
+  				backup_file $SYSTEM_EXT/$C
     		  fi
     	  done
 	  fi  
@@ -129,9 +93,9 @@ case "$1" in
   	print "Restoring LiteGapps"
   	if [ -f $base/list_install_system ]; then
   		for A in $(cat $base/list_install_system); do
-  			dir1=`dirname $SYSTEM/$A`
-  			sedlog "  Restoring •> $SYSTEM/$A"
-  			RESTORE_FILE $SYSTEM/$A
+  			dir1=`dirname $S/$A`
+  			sedlog "  Restoring •> $S/$A"
+  			restore_file $S/$A
   			ch_con $dir1
     	  done
   	fi
@@ -139,7 +103,7 @@ case "$1" in
   		for B in $(cat $base/list_install_product); do
   			dir1=`dirname $PRODUCT/$B`
   			sedlog "  Restoring •> $PRODUCT/$B"
-  			RESTORE_FILE $PRODUCT/$B
+  			restore_file $PRODUCT/$B
   			ch_con $dir1
     	  done
   	fi
@@ -147,7 +111,7 @@ case "$1" in
   		for C in $(cat $base/list_install_system_ext); do
   			dir1=`dirname $SYSTEM_EXT/$C`
   			sedlog "  Restoring •> $SYSTEM_EXT/$C"
-  			RESTORE_FILE $SYSTEM_EXT/$C
+  			restore_file $SYSTEM_EXT/$C
   			ch_con $dir1
     	  done
   	fi
@@ -165,7 +129,6 @@ case "$1" in
   pre-restore)
     # Stub
     print "Litegapps addon.d"
-    rm -rf $DIR
   ;;
   post-restore)
     echo " " >> $log
